+++ src/containers/EditListingPage/EditListingWizard/EditListingPricingPanel/EditListingPricingPanel.js
@@ -5,8 +5,9 @@ import classNames from 'classnames';
 import { FormattedMessage } from '../../../../util/reactIntl';
 import { LISTING_STATE_DRAFT, propTypes } from '../../../../util/types';
 import { types as sdkTypes } from '../../../../util/sdkLoader';
+import { isPriceVariationsEnabled } from '../../../../util/configHelpers';
 import { isValidCurrencyForTransactionProcess } from '../../../../util/fieldHelpers';
-import { FIXED } from '../../../../transactions/transaction';
+import { FIXED, isBookingProcess } from '../../../../transactions/transaction';
 
 // Import shared components
 import { H3, ListingLink } from '../../../../components';
@@ -34,14 +35,25 @@ const getListingTypeConfig = (publicData, listingTypes) => {
 // exporting helper functions that handle the initial values and the submission values.
 // This is a tentative approach to contain logic in one place.
 const getInitialValues = props => {
-  const { listing } = props;
-  const { unitType } = listing?.attributes?.publicData || {};
-  return unitType === FIXED
-    ? { ...getInitialValuesForPriceVariants(props), ...getInitialValuesForStartTimeInterval(props) }
+  const { listing, listingTypes } = props;
+  const { publicData } = listing?.attributes || {};
+  const { unitType } = publicData || {};
+  const listingTypeConfig = getListingTypeConfig(publicData, listingTypes);
+  // Note: publicData contains priceVariationsEnabled if listing is created with priceVariations enabled.
+  const isPriceVariationsInUse = isPriceVariationsEnabled(publicData, listingTypeConfig);
+
+  return unitType === FIXED || isPriceVariationsInUse
+    ? {
+        ...getInitialValuesForPriceVariants(props, isPriceVariationsInUse),
+        ...getInitialValuesForStartTimeInterval(props),
+      }
     : { price: listing?.attributes?.price };
 };
 
-const getEstimatedListing = (listing, updateValues) => {
+// This is needed to show the listing's price consistently over XHR calls.
+// I.e. we don't change the API entity saved to Redux store.
+// Instead, we use a temporary entity inside the form's state.
+const getOptimisticListing = (listing, updateValues) => {
   const tmpListing = {
     ...listing,
     attributes: {
@@ -101,7 +113,12 @@ const EditListingPricingPanel = props => {
 
   const publicData = listing?.attributes?.publicData;
   const listingTypeConfig = getListingTypeConfig(publicData, listingTypes);
-  const transactionProcessAlias = listingTypeConfig.transactionType.alias;
+  const transactionProcessAlias = listingTypeConfig?.transactionType?.alias;
+  const process = listingTypeConfig?.transactionType?.process;
+  const isBooking = isBookingProcess(process);
+
+  // Note: publicData contains priceVariationsEnabled if listing is created with priceVariations enabled.
+  const isPriceVariationsInUse = isPriceVariationsEnabled(publicData, listingTypeConfig);
 
   const isCompatibleCurrency = isValidCurrencyForTransactionProcess(
     transactionProcessAlias,
@@ -140,8 +157,8 @@ const EditListingPricingPanel = props => {
             // New values for listing attributes
             let updateValues = {};
 
-            if (unitType === FIXED) {
-              let publicDataUpdates = {};
+            if (unitType === FIXED || isPriceVariationsInUse) {
+              let publicDataUpdates = { priceVariationsEnabled: isPriceVariationsInUse };
               // NOTE: components that handle price variants and start time interval are currently
               // exporting helper functions that handle the initial values and the submission values.
               // This is a tentative approach to contain logic in one place.
@@ -156,31 +173,43 @@ const EditListingPricingPanel = props => {
               const priceVariantChanges = handleSubmitValuesForPriceVariants(
                 values,
                 publicDataUpdates,
-                unitType
+                unitType,
+                listingTypeConfig
               );
               updateValues = {
                 ...priceVariantChanges,
                 ...startTimeIntervalChanges,
                 publicData: {
+                  priceVariationsEnabled: isPriceVariationsInUse,
                   ...startTimeIntervalChanges.publicData,
                   ...priceVariantChanges.publicData,
                 },
               };
             } else {
-              updateValues = { price };
+              const priceVariationsEnabledMaybe = isBooking
+                ? {
+                    publicData: {
+                      priceVariationsEnabled: false,
+                    },
+                  }
+                : {};
+              updateValues = { price, ...priceVariationsEnabledMaybe };
             }
 
             // Save the initialValues to state
             // Otherwise, re-rendering would overwrite the values during XHR call.
             setState({
               initialValues: getInitialValues({
-                listing: getEstimatedListing(listing, updateValues),
+                listing: getOptimisticListing(listing, updateValues),
+                listingTypes,
               }),
             });
             onSubmit(updateValues);
           }}
           marketplaceCurrency={marketplaceCurrency}
           unitType={unitType}
+          listingTypeConfig={listingTypeConfig}
+          isPriceVariationsInUse={isPriceVariationsInUse}
           listingMinimumPriceSubUnits={listingMinimumPriceSubUnits}
           saveActionMsg={submitButtonText}
           disabled={disabled}