+++ src/containers/SearchPage/SearchPage.test.js
@@ -80,10 +80,6 @@ const listingFields = [
     // Formerly used for category, but now there's dedicated category setup
     key: 'cat',
     scope: 'public',
-    listingTypeConfig: {
-      limitToListingTypeIds: true,
-      listingTypeIds: ['sell-bicycles'],
-    },
     categoryConfig: {
       limitToCategoryIds: true,
       categoryIds: ['cats'],
@@ -103,12 +99,29 @@ const listingFields = [
     },
   },
   {
-    key: 'singleSelectTest',
+    key: 'boat ',
     scope: 'public',
     listingTypeConfig: {
       limitToListingTypeIds: true,
       listingTypeIds: ['sell-bicycles'],
     },
+    schemaType: 'enum',
+    enumOptions: [{ option: 'boat_1', label: 'Boat 1' }, { option: 'boat_2', label: 'Boat 2' }],
+    filterConfig: {
+      indexForSearch: true,
+      label: 'Boat',
+      group: 'primary',
+    },
+    showConfig: {
+      label: 'Boat',
+    },
+    saveConfig: {
+      label: 'Boat',
+    },
+  },
+  {
+    key: 'singleSelectTest',
+    scope: 'public',
     schemaType: 'enum',
     enumOptions: [{ option: 'enum1', label: 'Enum 1' }, { option: 'enum2', label: 'Enum 2' }],
     filterConfig: {
@@ -127,10 +140,6 @@ const listingFields = [
   {
     key: 'amenities',
     scope: 'public',
-    listingTypeConfig: {
-      limitToListingTypeIds: true,
-      listingTypeIds: ['rent-bicycles-daily', 'rent-bicycles-nightly', 'rent-bicycles-hourly'],
-    },
     schemaType: 'multi-enum',
     enumOptions: [{ option: 'dog_1', label: 'Dog 1' }, { option: 'dog_2', label: 'Dog 2' }],
     filterConfig: {
@@ -148,6 +157,8 @@ const listingFields = [
   },
 ];
 
+const listingTypeOptions = listingTypes.map(lt => ({ option: lt.id, label: labelize(lt.id) }));
+
 const defaultFiltersConfig = [
   {
     key: 'categoryLevel',
@@ -155,6 +166,12 @@ const defaultFiltersConfig = [
     scope: 'public',
     isNestedEnum: true,
     nestedParams: ['categoryLevel1', 'categoryLevel2', 'categoryLevel3'],
+  },
+  {
+    key: 'listingType',
+    schemaType: 'listingType',
+    scope: 'public',
+    options: listingTypeOptions,
   },
   {
     key: 'price',
@@ -269,6 +286,8 @@ const getSearchParams = config => {
       // when transitioning from search page to listing page
       'publicData.pickupEnabled',
       'publicData.shippingEnabled',
+      'publicData.priceVariationsEnabled',
+      'publicData.priceVariants',
     ],
     'fields.user': ['profile.displayName', 'profile.abbreviatedName'],
     'fields.image': [
@@ -324,6 +343,8 @@ describe('SearchPage', () => {
 
       // Has no Cat filter (primary filter tied to 'Cats' category)
       expect(queryByText('Cat')).not.toBeInTheDocument();
+      // Has no Boat filter (primary filter tied to 'sell-bicycles' listing type)
+      expect(queryByText('Boat')).not.toBeInTheDocument();
       // Has(!) Amenities filter (secondary filter)
       expect(getByText('Amenities')).toBeInTheDocument();
       // Has Single Select Test filter
@@ -340,6 +361,13 @@ describe('SearchPage', () => {
       expect(getByText('Fish')).toBeInTheDocument();
       expect(queryByText('Freshwater')).not.toBeInTheDocument();
 
+      // Has Listing type filter
+      expect(getByText('FilterComponent.listingTypeLabel')).toBeInTheDocument();
+      expect(getByText('Rent bicycles daily')).toBeInTheDocument();
+      expect(getByText('Rent bicycles nightly')).toBeInTheDocument();
+      expect(getByText('Rent bicycles hourly')).toBeInTheDocument();
+      expect(getByText('Sell bicycles')).toBeInTheDocument();
+
       // Has Price filter
       expect(getByText('FilterComponent.priceLabel')).toBeInTheDocument();
 
@@ -349,7 +377,7 @@ describe('SearchPage', () => {
       // Has listing with title
       expect(getByText('l2 title')).toBeInTheDocument();
       // 2 listings with the same price
-      expect(getAllByText('$55.00')).toHaveLength(2);
+      expect(getAllByText('ListingCard.price')).toHaveLength(2);
     });
 
     // Test category intercation: click "Fish"
@@ -414,6 +442,13 @@ describe('SearchPage', () => {
       expect(queryByText('Cats')).not.toBeInTheDocument();
       expect(queryByText('Fish')).not.toBeInTheDocument();
 
+      // Has Listing type filter
+      expect(getByText('FilterComponent.listingTypeLabel')).toBeInTheDocument();
+      expect(queryByText('Rent bicycles daily')).not.toBeInTheDocument();
+      expect(queryByText('Rent bicycles nightly')).not.toBeInTheDocument();
+      expect(queryByText('Rent bicycles hourly')).not.toBeInTheDocument();
+      expect(queryByText('Sell bicycles')).not.toBeInTheDocument();
+
       // Has "more filters" button for secondary filters
       expect(getByText('SearchFiltersPrimary.moreFiltersButton')).toBeInTheDocument();
 
@@ -426,7 +461,7 @@ describe('SearchPage', () => {
       // Has listing with title
       expect(getByText('l2 title')).toBeInTheDocument();
       // 2 listings with the same price
-      expect(getAllByText('$55.00')).toHaveLength(2);
+      expect(getAllByText('ListingCard.price')).toHaveLength(2);
     });
 
     // Test category intercation
@@ -502,6 +537,73 @@ describe('SearchPage', () => {
     expect(getByText('Fish')).toBeInTheDocument();
     expect(queryByText('Freshwater')).not.toBeInTheDocument();
     expect(queryByText('Saltwater')).not.toBeInTheDocument();
+  });
+
+  it('Check that Boat filters is revealed in grid variant', async () => {
+    // Select correct SearchPage variant according to route configuration
+    const config = getConfig('grid');
+    const routeConfiguration = getRouteConfiguration(config.layout);
+    const props = { ...commonProps };
+    const searchRouteConfig = routeConfiguration.find(conf => conf.name === 'SearchPage');
+    const SearchPage = searchRouteConfig.component;
+
+    const { getByPlaceholderText, getByText, getAllByText, queryByText, getByRole } = render(
+      <SearchPage {...props} />,
+      {
+        initialState,
+        config,
+        routeConfiguration,
+      }
+    );
+
+    await waitFor(() => {
+      // Has no Boat filter (primary)
+      expect(queryByText('Boat')).not.toBeInTheDocument();
+
+      // Has Listing type filter
+      expect(getByText('FilterComponent.listingTypeLabel')).toBeInTheDocument();
+      expect(getByText('Rent bicycles daily')).toBeInTheDocument();
+      expect(getByText('Rent bicycles nightly')).toBeInTheDocument();
+      expect(getByText('Rent bicycles hourly')).toBeInTheDocument();
+      expect(getByText('Sell bicycles')).toBeInTheDocument();
+    });
+
+    // Test category intercation: click "Sell bicycles"
+    await waitFor(() => {
+      userEvent.click(getByRole('button', { name: 'Sell bicycles' }));
+    });
+
+    // Has Boat filter filter (primary)
+    expect(getByText('Boat')).toBeInTheDocument();
+  });
+
+  it('Check that Listing type filter is not revealed when using a listing type path param', async () => {
+    // Select correct SearchPage variant according to route configuration
+    const config = getConfig('grid');
+    const routeConfiguration = getRouteConfiguration(config.layout);
+    const props = { ...commonProps, params: { listingType: 'sell-bicycles' } };
+    const searchRouteConfig = routeConfiguration.find(
+      conf => conf.name === 'SearchPageWithListingType'
+    );
+    const SearchPage = searchRouteConfig.component;
+
+    const { getByPlaceholderText, getByText, getAllByText, queryByText, getByRole } = render(
+      <SearchPage {...props} />,
+      {
+        initialState,
+        config,
+        routeConfiguration,
+      }
+    );
+
+    await waitFor(() => {
+      // Does not have Listing type filter
+      expect(queryByText('FilterComponent.listingTypeLabel')).not.toBeInTheDocument();
+      expect(queryByText('Rent bicycles daily')).not.toBeInTheDocument();
+      expect(queryByText('Rent bicycles nightly')).not.toBeInTheDocument();
+      expect(queryByText('Rent bicycles hourly')).not.toBeInTheDocument();
+      expect(queryByText('Sell bicycles')).not.toBeInTheDocument();
+    });
   });
 });